requirements.txt
robotframework==6.1.1
robotframework-requests==0.9.5
robotframework-browser==17.5.2
robotframework-jsonlibrary==0.5
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2